Move tests to plugin test folder, and fix the syntax errors.
