+++ plugins/stack-snippets/src/stackSnippetPlugin.ts
@@ -71,74 +71,80 @@ const snippetOpts: StackSnippetOptions = {
     }
 };
 
-export const stackSnippetPlugin: EditorPlugin = () => ({
-    richText: {
-        nodeViews: {
-            stack_snippet: (
-                node: ProseMirrorNode,
-                view: EditorView,
-                getPos: () => number
-            ) => {
-                return new StackSnippetView(
-                    node,
-                    view,
-                    getPos,
-                    snippetOpts
-                );
-            }
-        },
-        plugins: [
-            stackSnippetPasteHandler
-        ]
-    },
-    extendSchema: (schema) => {
-        schema.nodes = schema.nodes.append(stackSnippetRichTextNodeSpec);
-        return schema;
-    },
-    markdown: {
-        parser: {
-            ...stackSnippetMarkdownParser
-        },
-        serializers: {
-            nodes: {
-                ...stackSnippetMarkdownSerializer
+/**
+ * Build the StackSnippet plugin using hoisted options that can be specified at runtime
+ *
+ * This allows callers to specify how to handle rendering and modal opening of buttons externally to the editor.
+ * **/
+export const stackSnippetPlugin: (opts?: StackSnippetOptions) => EditorPlugin = (opts) =>
+    () => ({
+        richText: {
+            nodeViews: {
+                stack_snippet: (
+                    node: ProseMirrorNode,
+                    view: EditorView,
+                    getPos: () => number
+                ) => {
+                    return new StackSnippetView(
+                        node,
+                        view,
+                        getPos,
+                        opts
+                    );
+                }
             },
-            marks: {},
+            plugins: [
+                stackSnippetPasteHandler
+            ]
         },
-        alterMarkdownIt: (mdit) => {
-            mdit.use(markdownPlugin);
-        }
-    },
-    menuItems: (schema, coreMenus) => {
-        const clone = coreMenus.find(mb => mb.name == "code-formatting");
-
-        return [
-            {
-                ...clone,
-                entries: [
-                    {
-                        key: "openSnippetModal",
-                        richText: {
-                            command: openSnippetModal(snippetOpts),
-                        },
-                        commonmark: {
-                            command: openSnippetModal(snippetOpts),
-                        },
-                        display: makeMenuButton(
-                            "Play",
-                            {
-                                title: _t("commands.stack_snippet.title", {
-                                    shortcut: getShortcut("Mod-9"),
-                                }),
-                                description: _t(
-                                    "commands.stack_snippet.description"
-                                ),
-                            },
-                            "stack-snippet-open-btn"
-                        ),
-                    }
-                ]
+        extendSchema: (schema) => {
+            schema.nodes = schema.nodes.append(stackSnippetRichTextNodeSpec);
+            return schema;
+        },
+        markdown: {
+            parser: {
+                ...stackSnippetMarkdownParser
+            },
+            serializers: {
+                nodes: {
+                    ...stackSnippetMarkdownSerializer
+                },
+                marks: {},
+            },
+            alterMarkdownIt: (mdit) => {
+                mdit.use(markdownPlugin);
             }
-        ]
-    }
-});
+        },
+        menuItems: (schema, coreMenus) => {
+            const clone = coreMenus.find(mb => mb.name == "code-formatting");
+
+            return [
+                {
+                    ...clone,
+                    entries: [
+                        {
+                            key: "openSnippetModal",
+                            richText: {
+                                command: openSnippetModal(opts),
+                            },
+                            commonmark: {
+                                command: openSnippetModal(opts),
+                            },
+                            display: makeMenuButton(
+                                "Play",
+                                {
+                                    title: _t("commands.stack_snippet.title", {
+                                        shortcut: getShortcut("Mod-9"),
+                                    }),
+                                    description: _t(
+                                        "commands.stack_snippet.description"
+                                    ),
+                                },
+                                "stack-snippet-open-btn"
+                            ),
+                        }
+                    ]
+                }
+            ]
+        }
+    });
