+++ test/rich-text/editor.test.ts
@@ -1,10 +1,6 @@
 import { RichTextEditor, RichTextOptions } from "../../src/rich-text/editor";
 import * as mdp from "../../src/shared/markdown-parser";
 import { externalPluginProvider, normalize } from "../test-helpers";
-import {
-    validSnippetRenderCases,
-    invalidSnippetRenderCases,
-} from "./plugins/stack-snippet-helpers";
 
 // mock the markdown-parser for testing
 jest.mock("../../src/shared/markdown-parser");
@@ -323,137 +319,6 @@ _world_.
                 "right"
             );
         });
-    });
-
-    describe("stack snippets", () => {
-        //Valid snippets
-        // - Just JS
-        // - Just CSS
-        // - Just HTML
-        // any two of the three
-        // all three
-        // Special cases:
-        // if the HTML contains comments that look like snippets code
-
-        const shouldHaveSnippetBlock = (rendered: Element) => {
-            expect(rendered.querySelectorAll("div.snippet")).toHaveLength(1);
-            expect(rendered.querySelectorAll("div.snippet-code")).toHaveLength(
-                1
-            );
-            expect(
-                rendered.querySelectorAll("div.snippet-result")
-            ).toHaveLength(1);
-            expect(
-                rendered.querySelectorAll("div.snippet-result")
-            ).toHaveLength(1);
-            const resultCode = rendered.querySelectorAll(
-                "div.snippet-result-code"
-            );
-            expect(resultCode).toHaveLength(1);
-            expect(resultCode[0].childNodes).toHaveLength(0);
-        };
-
-        const shouldHaveRunCodeButton = (rendered: Element) => {
-            const runCode = rendered.querySelectorAll(
-                "div.snippet-ctas > button"
-            );
-            expect(runCode).toHaveLength(1);
-            expect(runCode[0].attributes.getNamedItem("title").value).toBe(
-                "Run code snippet"
-            );
-            expect(runCode[0].attributes.getNamedItem("aria-label").value).toBe(
-                "Run code snippet"
-            );
-        };
-
-        const shouldHaveLanguageBlocks = (
-            rendered: Element,
-            langs: string[]
-        ) => {
-            for (const lang of langs) {
-                const langBlock = rendered.querySelectorAll(
-                    `div.snippet-code > pre.lang-${lang} > code`
-                );
-                expect(langBlock).toHaveLength(1);
-                expect(langBlock[0].innerHTML).toBeTruthy();
-            }
-
-            const codeBlocks = rendered.querySelectorAll(
-                `div.snippet-code > pre`
-            );
-            expect(codeBlocks).toHaveLength(langs.length);
-        };
-
-        it.each(validSnippetRenderCases)(
-            "should render snippets",
-            (markdown: string, langs: string[]) => {
-                const richEditorView = richView(markdown, {
-                    stackSnippet: {
-                        renderer: () => null,
-                        openSnippetsModal: () => {},
-                    },
-                });
-
-                const rendered = richEditorView.dom;
-                shouldHaveSnippetBlock(rendered);
-                shouldHaveRunCodeButton(rendered);
-                shouldHaveLanguageBlocks(rendered, langs);
-            }
-        );
-
-        it.each(validSnippetRenderCases)(
-            "should render without button if no render supplied",
-            (markdown: string, langs: string[]) => {
-                const richEditorView = richView(markdown, {});
-
-                const rendered = richEditorView.dom;
-                shouldHaveSnippetBlock(rendered);
-                shouldHaveLanguageBlocks(rendered, langs);
-            }
-        );
-
-        //Snippets with four spaces should render as if in a code block
-        it("should render four-spaced snippets within a code block", () => {
-            const markdown = `    <!-- begin snippet: js hide: false console: true babel: null babelPresetReact: false babelPresetTS: false -->
-
-    <!-- language: lang-js -->
-
-        console.log("test");
-
-    <!-- end snippet -->`;
-
-            const richEditorView = richView(markdown, {
-                stackSnippet: {
-                    renderer: () => null,
-                    openSnippetsModal: () => {},
-                },
-            });
-
-            const rendered = richEditorView.dom;
-            const codeBlocks = rendered.querySelectorAll(
-                "div > pre.s-code-block > code"
-            );
-            expect(codeBlocks).toHaveLength(1);
-            expect(codeBlocks[0].innerHTML).toBeTruthy();
-        });
-
-        it.each(invalidSnippetRenderCases)(
-            "should not render invalid snippets",
-            (markdown) => {
-                const richEditorView = richView(markdown, {
-                    stackSnippet: {
-                        renderer: () => null,
-                        openSnippetsModal: () => {},
-                    },
-                });
-
-                const rendered = richEditorView.dom;
-
-                expect(rendered.querySelectorAll("div.snippet")).toHaveLength(
-                    0
-                );
-            }
-        );
     });
 
     describe("general", () => {
